Extract `Column` component
